refactor: moves dela APIs to a separate module
- no code changes
